Finish up the ability to upload any RestTemplate-supported Resource to WordPress.
In short, if a resource is local and needs to be uploaded, it will do the required
gymnastics that conforms to web standards for uploading. If a resource is remote,
and it is supported by UrlResource (Spring), it will be sent to WordPress which will
then pull the resource from there (depending on the visibility of that URL to
WordPress).